chore: cleanup utils
