+++ utils/titleCase.ts
@@ -1,7 +1,6 @@
-export default function titleCase(str: string) {
-  var splitStr = str.toLowerCase().split(' ')
-  for (var i = 0; i < splitStr.length; i++) {
-    splitStr[i] = splitStr[i].charAt(0).toUpperCase() + splitStr[i].substring(1)
-  }
-  return splitStr.join(' ')
-}
+/**
+ * Converts text to Title Case.
+ */
+const titleCase = (str: string) => str.replace(/\b\S/g, (t) => t.toUpperCase())
+
+export default titleCase
